Enable -Wdouble-promotion and fix all warnings
I recently read this excellent article that, among other things,
recommended I enable this warning. It's useful to know when floats are
being implicitly promoted, so I'm leaving this on.

This also caught a few seemingly accidental promotions in stb_image.h. I'm
using an old version, but I'll have to see if these should be
upstreamed.

Article link:
https://nullprogram.com/blog/2023/04/29/

+++ src/datatypes/color.c
@@ -97,5 +97,5 @@ struct color color_from_hsl(float hue, float saturation, float lightness) {
 }
 
 void color_dump(struct color c, char *buf, size_t bufsize) {
-	snprintf(buf, bufsize, "{ %.2f, %.2f, %.2f, %.2f }", c.red, c.green, c.blue, c.alpha);
+	snprintf(buf, bufsize, "{ %.2f, %.2f, %.2f, %.2f }", (double)c.red, (double)c.green, (double)c.blue, (double)c.alpha);
 }
