+++ src/utils/protocol/worker.c
@@ -271,9 +271,9 @@ static cJSON *startRender(int connectionSocket) {
 			}
 			cJSON *stats = newAction("stats");
 			cJSON_AddNumberToObject(stats, "completed", completedSamples);
-			cJSON_AddNumberToObject(stats, "avgPerPass", avgTimePerTilePass);
+			cJSON_AddNumberToObject(stats, "avgPerPass", (double)avgTimePerTilePass);
 			mutex_lock(g_worker_socket_mutex);
-			logr(debug, "Sending stats update for: %"PRIu64", %.2f\n", completedSamples, avgTimePerTilePass);
+			logr(debug, "Sending stats update for: %"PRIu64", %.2f\n", completedSamples, (double)avgTimePerTilePass);
 			if (!sendJSON(connectionSocket, stats, NULL)) {
 				logr(debug, "Connection lost, bailing out.\n");
 				// Setting this flag also kills the threads.
